messages added

+++ MuseumForm/ComboBoxItem.cs
@@ -9,7 +9,7 @@ namespace MuseumForm
     public class ComboboxItem
     {
         public string Text { get; set; }
-        public long Value { get; set; }
+        public int Value { get; set; }
         public double doubleValue { get; set; }
 
         public override string ToString()
